controllers in separate dirs, routes file organised, dash controller working with frontend using request.params + username, sending back the email, can now query db for any info we want, login + register tested too

+++ src/controllers/user/deleteUser.controller.ts
@@ -2,7 +2,6 @@ import { FastifyRequest, FastifyReply } from 'fastify';
 
 export async function deleteUser(request: FastifyRequest, reply: FastifyReply): Promise<void> {
 
-	// ensure user is logged in, then check credentials again
 	const {username, password} = request.body as {username: string, password: string};
 	try {
 		const db = request.server.db;
@@ -12,16 +11,22 @@ export async function deleteUser(request: FastifyRequest, reply: FastifyReply): 
 		}
 		const stmt = db.prepare('SELECT * FROM userTable WHERE username = ? AND password = ?');
 		const user = stmt.get(username, password);
-		if (user)
-		{
-			const stmt = db.prepare("DELETE FROM userTable WHERE username = ? AND password = ?");
-			const result = stmt.run(username, password);
-			return reply.send({ message: `User successfully deleted: ${username}` });
-			// pop up, account deleted, return to homepage
+		if (user) {
+			if (user.status == 1) {
+
+				const stmt = db.prepare("DELETE FROM userTable WHERE username = ? AND password = ?");
+				const result = stmt.run(username, password);
+				return reply.send({ message: `User successfully deleted: ${username}` });
+				// pop up, account deleted, return to homepage
+			}
+			else {
+				return reply.send({ error: "user status = not logged in" });
+				// return to delete page
+			}
 		}
 		else
 			reply.send({ error: "Invalid credentials / non-existent user" });
-		// refresh delete page
+			// refresh delete page
 	}
 	catch (err) {
 		console.log(err);
